Remove a lot of very old, obsolete content.
This documentation portal hasn't seen significant activity in years,
and it mostly documents a system that no longer exists. To avoid
confusion for developers in future, let's remove the docs that we
know to be outdated.

Broadly:

* Remove all the pre-Firefox-Accounts sync auth infrastructure
  (e.g. reg, sreg) since we haven't run that in many years.
* Remove services that supported Firefox OS and related efforts
  (e.g. msisdn-gateway, loop) since they've long since been
  decommissioned.
* Remove things that happened to live here because they were
  developed by the then-Services team (e.g. Heka, FHR) and have
  since been either retired or moved.
* Remove the entirely-hypothetical "Storage Format 6" from the
  sync client docs.

This leaves us with a much leaner docs site describing the sync
client and its supporting servers. I think it would be useful
to pull those docs closer the code that they document over time,
but this is at least a step in a healthy direction.